refactor: move edit button into separate file
