+++ src/livePreview/eventManager/livePreviewEventManger.constant.ts
@@ -2,6 +2,7 @@ export const LIVE_PREVIEW_POST_MESSAGE_EVENTS = {
     INIT: "init",
     ON_CHANGE: "client-data-send",
     HISTORY: "history",
+    CHECK_ENTRY_PAGE: "check-entry-page",
 } as const;
 
 export const LIVE_PREVIEW_CHANNEL_ID = "live-preview";
