+++ src/livePreview/types/onEntryChangeCallback.type.ts
@@ -1,0 +1,16 @@
+export declare type OnEntryChangeCallback = () => void;
+
+export declare type OnEntryChangeConfig = {
+    skipInitialRender?: boolean;
+};
+
+export declare type OnEntryChangeCallbackUID = string;
+
+export type OnEntryChangeCallbackSubscribers = Record<
+    OnEntryChangeCallbackUID,
+    OnEntryChangeCallback
+>;
+
+export type OnEntryChangeUnsubscribeParameters =
+    | OnEntryChangeCallbackUID
+    | OnEntryChangeCallback;
